ostream operator code
we need to have a counter that keeps track of how many numbers there are total in the list. b/c we need to output all the values in the list 4 per column.We added code for the ostream << operator. and we worked on the output by moving the current and couting the data. i am not sure if this is right, but we gave it a good shot.